+++ imaging/vtkXTextMapper.cxx
@@ -320,7 +320,7 @@ void vtkXTextMapper::RenderOverlay(vtkViewport* viewport, vtkActor2D* actor)
   // adjust actorPos to account for justification
   int pos[2];
   pos[0] = actorPos[0];
-  pos[1] = actorPos[1] + this->LineOffset;
+  pos[1] = (int)(actorPos[1] + this->LineOffset);
   switch (this->Justification)
     {
     // do nothing for case 0 left
